Fixed a bunch of compiler warnings in XCode. Also "fixed" a linker warning by disabling it (in debug mode at least). The correct fix would be to recompile freeimage with -fvisibility=public or default.

+++ Gwen/Source/Gwen/Controls/Base.cpp
@@ -321,7 +321,7 @@ unsigned int Base::NumChildren()
 {
 	// Include m_InnerPanel's children here?
 
-	return Children.size();
+	return static_cast<int>(Children.size());
 }
 
 Controls::Base* Base::GetChild( unsigned int i )
